+++ mempool/blockchain.go
@@ -86,6 +86,11 @@ func (b Blockchain) CurrentBlock() *types.Header {
 	}
 
 	blockHeight := ctx.BlockHeight()
+	// prevent the reorg from triggering after a restart since previousHeaderHash is stored as an in-memory variable
+	if blockHeight > 1 && b.previousHeaderHash == (common.Hash{}) {
+		return b.zeroHeader
+	}
+
 	blockTime := ctx.BlockTime().Unix()
 	gasUsed := b.feeMarketKeeper.GetBlockGasWanted(ctx)
 	appHash := common.BytesToHash(ctx.BlockHeader().AppHash)